+++ client/app/queue/pulacCerullo/PulacCerulloReminderAlert.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Alert from '../components/Alert';
+import Alert from '../../components/Alert';
 import { chairmanMemoUrl } from '.';
 import { CavcLinkInfo } from './CavcLinkInfo';
